Update and fixed url parameter in update_ruleset
Before, the update_ruleset will upgrade the local ruleset if the major versions of the ruleset were equal. Now this script will upgrade your ruleset only if the major and minor versions are equal. Also, now, the url parameter can compute the branch parameter from the url if it is missing.