0090-telnetd_rules.xml updated
We fixed a bug where events that were "decoded" by the telnetd decoder did not access these rules, due to the syslog format.

We have also fixed a bug where rule 5603 was never activated, as it did not come from any other rule, it could not successfully check the content of the <match> field.